Refactor views and add color extension
