+++ PetTranslator/PetTranslator/Models/SettingsItem.swift
@@ -1,0 +1,13 @@
+//
+//  SettingsItem.swift
+//  PetTranslator
+//
+//  Created by Keto Nioradze on 18.04.25.
+//
+
+import Foundation
+
+struct SettingsItem: Identifiable {
+    let id = UUID()
+    let title: String
+}
