+++ PetTranslator/PetTranslator/Views/ResultsView.swift
@@ -55,7 +55,7 @@ struct ResultsView: View {
                             .resizable()
                             .frame(width: 295, height: 248)
 
-                        Text(getRandomText(for: selectedPet))
+                        Text(PetTranslatorModel.getRandomMessage(for: selectedPet))
                             .font(.custom("Konkhmer Sleokchher", size: 16))
                             .foregroundColor(Color(hex: "#292D32"))
                             .padding(.top, 60)
@@ -64,40 +64,6 @@ struct ResultsView: View {
                 }
             }
             .padding(.top, 100)
-        }
-    }
-
-    // MARK: - Helpers
-    func getRandomText(for pet: String) -> String {
-        if pet == "dog" {
-            let messages = [
-                "I’m hungry, feed me!",
-                "What are you doing, human?",
-                "Is it walk time yet?",
-                "I smelled something amazing!",
-                "You’re my favorite human.",
-                "Can I have that snack now?",
-                "I studied barkology at Pawvard.",
-                "I did not chew the slipper, I swear!",
-                "Throw the ball already!",
-                "Sniff sniff… smells like adventure!",
-                "Cats are real or government spies?",
-                "You smell like love and peanut butter."
-            ]
-            return messages.randomElement() ?? "Processing..."
-        } else {
-            let messages = [
-                "I’m hungry, feed me!",
-                "What are you doing, human?",
-                "I was napping… what is it now?",
-                "Feed me. Immediately.",
-                "I knocked it down on purpose.",
-                "I’m judging you.",
-                "That thing you love? I chewed it.",
-                "Purr activated. You're welcome.",
-                "Your laptop is warm. I claim it."
-            ]
-            return messages.randomElement() ?? "Processing..."
         }
     }
 }
